feat(wasm): update pkg to build with webpack and produce isomorphic dist for node and browsers
