+++ wasm/pkg/.gitignore
@@ -1,1 +1,4 @@
+node_modules/
+dist/
+package-lock.json
 wasm_exec.js
